Simplify experiment segment name validation
Instead of using a regular expression to check for invalid characters,
we use a simple for-loop.  This allows us to get rid of the regex
object, which is the only thing in the .bss section and thus we are
saving 4KB of memory.

Also fix a bug in the validation logic where previously "foo.bar."
would be deemed valid.

Bug: 138856262
Test: Run unit tests.
Test: objdump -h server_configurable_flags.so. Check .bss is gone.
Change-Id: Id9bed9bed7c2cc5b2e4bdb8f8e2e9b2611ddbe2f